backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation and processing
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.4

# HTTP requests
requests==2.31.0

# File handling and CSV processing
python-csv==0.0.13

# Security and environment
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
uuid==1.30

# Development (optional)
pytest==7.4.4

# --- eBay API ---
ebaysdk==2.2.0
authlib==1.3.0